Refactoring and cleanup of Project.pm
Commenting and refactoring when opportunity arises, move misplaced code
to its correct place in Project.pm or into AM.pm. Sum initialization
has nothing to do with the project files, so move it into AM.pm. The
_compute_vars method seems mostly concerned with how to divide the
variables among the four lattices, so move the code setting the data
vector length and the variable format out of it.