more modular accessors for outcomelist
Add more accessors to the information in outcomelist, but instead of
just returning the whole structure, provide 1) the size of the array
(or number of outcomes) and 2) the outcome at a given index. Refactor
AM.pm to use these wherever possible. _outcomelist cannot be removed
because the whole structure must be available to the hook methods.