test hook variables with outcome file
Create tests for the values of @outcomelist and %outcometonum within
hook methods when an outcome file is present.

@outcomelis and %outcometonum both differ when an actual outcome file
is present in the project. This is because both of these variables
relate to the "long" version of an outcome, which is only available in
a project outcome file. Without the outcome file, the "short" outcomes
are copied from the data set to replace the "long" outcomes.