make outcomelist consistent with or without outcome file
Previously, having an outcome file would result in a number of outcomes
equal to the number of data items, which made the number less useful
and probably messed up the printing of gangs. The documentation in
AM.pm claims that @outcomelist contains unique outcomes, but this is
not the case with an outcome file. Change the way outcomelist is
created when reading a project file so that it contains only unique
outcomes. Add a test for the return value of num_outcomes when using
an outcome file, and change tests where needed to expect unique
outcomes instead of the expanded list.