refactor all state change into _add_data
Previously, several aspects of project state were changed only after
the entirety of the project files had been read. Move all state changes
into the _add_data method so that it is kept updated every time a new
exemplar is added to the set. Change the format methods so that they
are read-only and use current state values to construct a format
variable when called (a very fast operation).

Along the way, remove now-unneeded methods related to outcome file
loading, and update TODO.txt as ideas occur.

With all of the state change now in one place, _add_data can safely be
made into a public method so that we no longer need to place the data
into files. This is a much more flexible solution.