pass package and version to XSLoader::load
Apparently not passing package and version to XSLoader is sort of a
newer feature, so for greater backwards-compatibility (and less fail
reports), add the two arguments in there.

Change the dzil package plugin so that there can be a version number in
the original file that is replaced when the build is made. This is
necessary for the XSLoader to work during development.