add `get_project` method to AM
Then test this accessor (and leave room for future accessors) in a new
test file. Move it and 01-classify.t into a new test folder for testing
the AM class.

Remove tests involving reading projects from disk from 01-classify.t,
which are no longer needed since it is verified that a project object
is correctly created.