`print_config` method now only needs one variable
Change `print_summary` to `print_config` (since it just prints
configuration information, not classification information). Refactor
all relevant configuration data to be in the `$result` variable, and
pass only that to the method (which is now a function). Rename
`$self->{eg}` to `$result{given_excluded}` so its function is no longer
a mystery.