add Result class
Write a new class, `Result`, that for now just holds all of the
configuration information for a single classification and can
print it. Use this new method and get rid of `print_config`, which
is no longer needed.

Needs tests, and maybe input checking (filed a request for that:
https://github.com/dagolden/Class-Tiny/issues/14). Also needs
documentation on accessors, but wait for those as some things might
change.