move statistical summary into Result
Move the statistical summary logic/printing into Result and call
that from AM.pm. This required the addition of several new accessors,
and Result now stores whether the predicted outcome was a tie or not.
This is not yet printed or tested anywhere, but will be important in
the future.

This is quite a lot of refactoring and will doubtless need more
refactoring later, as much of it still seems quite messy.