add tests/documentation for Result.pm
Add rough tests for `Result` that only test it through its printing
methods. This will suffice for now until further functionality is
added to Result. Also add documentation explaining the string info
methods, but not giving an exact format as I do not want anyone to
depend on it programmatically.