create an analogical set API and test through it
Add a method to `Result` to return the analogical set. Refactor
`Result` so that the new method and the printing method do not
duplicate work, instead retrieving and caching analogical set
information in a separate method. Change the printing method back
to simply outputting items in order of appearance in the data set;
this makes the cached structure easier to work with, since I don't
want the new method to return groups of contexts (which is the job
the gang effects reporting, and is not simple enough for the new
method, which already requires the user to lookup the data item
themselves).

Change the analogical set test (previously in `07-skipset.t`) to test
through this API instead, and move it into a method in the file with
the other classification tests.

Update TODO.