specify excluded data in Result
Store the indices of the excluded data items in `Result` instead of
simply the number of items that were excluded. Update tests for
`Result` to reflect this. Update the datahooks test to use the `Result`
API instead of reading printouts. Update all code in both files
to use the latest APIs instead of reading data from disk.

Update TODO.