add tie result to statistical_summary printout
Instead of only reporting correct/incorrect in the statistical_summary,
also report ties. Add tests for incorrect and tie situations, and
reorganize the tests for `Result`. Rename the test file, as the only
things currently tested in that file are summary printouts. The rest
of it is tested in the classification tests.