add dataset_from_file function, update add_item
To replace the reading functionality of `Project`, add
`dataset_from_file` to `DataSet`. This time require the user to be
explicit about the input format instead of having a default with
logging. Make the function an export. Update `add_item` to also accept
an `Item` object, and take advantage of this in `dataset_from_file`.
Update tests.