better support for unknown features and classes
Add functionality for reading files that contain items with unknown
classes. Allow the user to decide what string signifies unknown, with
`UNK` as the default. Extend this customizable support to features, as
well, so that by default `UNK` can be used for both unknown classes and
unknown features. Unknown features are set to undef in `Item`, and this will replace the hardcoded use of `=` in `AM.pm`. Update tests.

Change the default item comment to be a comma separated list so that
unknown feature values are easier to see (trailing whitespace is
definitely not).