save unknown features as '' instead of undef
I ran into an interesting problem with null features being represented
by undef. The `_context_label` method in `AM` will get stuck in an
infinite loop with `exclude_nulls` turned on. This is because `$index`
will often extend past the end of the array, and these elements are
autovivified as `undef`. So `++$index while !defined $array[$index]` is
deadly.

Change the null variable representation to simply `''`, which is
convenient because it is already difficult (impossible?) to represent
it in supported file formats, and it avoids the above described
problem.