batch state summary and fix pass method
Create a method for printing the current state in `Batch`. During this
change it was discovered that `pass` was starting at 0, which threw off
if statements. Set pass to undef when nothing is currently being
processed, and when it is set then make it start at 1, not at 0.
Update tests accordingly.