redo hook parameters
Using methods for iteration-related variables did not work out. It was
complicated to implement, and would have required careful reading of
the documentation for users to get it right. Instead, pass all these
variables as parameters.

After understanding the organization of the hooks better, change
`endtesthook` so that it does not receive the iteration number,
which is irrelevant because of the repeat hooks, and to get all of the
result objects for the item (there are multiple if there are multiple
iterations).

As the iteration-specific information was removed from object state,
there is no longer any real point in the `state_summary` method, which
was meant to fill holes opened by the loss of this information in the
reports given by `Result`.

Do any other convenient refactoring/cleanup along the way, including
moving time calculations into the `if($log->is_info)` blocks, and
removing the unneeded `continue` block.