rename unknown to null; null is =, not UNK
I remembered that `null` doesn't just mean `unknown`, but also N/A
or non-existent, etc. So change the documentation and APIs back to
`null`, and add a separate option for specifying the null string,
since unknown class and null feature value are two different things.

While I'm at it, the null string should be `=` like it always was,
since that is only one character and can be used in the nocommas
format, whereas UNK cannot.