make Item immutable
I was worried about the kinds of changes that could be made to Item
after it had already been used in classification and thus stored in
a Result object. Changing it could really confuse someone trying to
interpret results printed later on. So make Item immutable, removing
the `Class::Tiny` usage and instead providing a constructor and
getters.

The removal of `Class::Tiny` somehow changed the place where `croak`
reports errors to what I expected it to be: the calling file. This
is fortunate, and means that the tests can verify correct behavior now.
Update tests accordingly.