re-indent AM.xs
The indenting of some very nested, very long loop structures was
reduced by a significant amount, and in some places indent was not
done, or done strangely. Reindent the entire file with 2 spaces.

Note that the comment on line 341 must not be un-indented; leaving
no indent prevents the file from compiling because of the XS
pre-processor misinterpreting it as a function declaration.