refactor active_feats
`active_feats` is a bad name, since it implies that it contains a list
of the features that are active. Change it to `lattice_sizes`, because
it contains the size in features of each lattice.

Change this variable so that it is passed into `_fillandcount`, instead
of being saved in `$self` and needing to be edited in place instead of
being assigned to.