combine most code in linear/quadratic blocks
Eliminate a massive amount of code duplication. The code for combining
sublattices and for counting occurrences was completely duplicated by
an if/else block check for the quadratic/linear setting. There are
about 25 lines that are needed only for quadratic counting, but the rest
are exactly the same for both blocks. Partitioning off 25 special lines
and then combine the rest. About 120 lines eliminated, and the program
is now much easier to follow.