Fix double `sizeof` call
XS's `Copy` requires as type as the 4th parameter and then likely calls
`sizeof` on it. All of the `Copy` calls except this one were correctly
passing in a type. This one called `sizeof` on the argument first, so
that the requested memory unit size indicated was always
`sizeof(size_t)`. This is usually 8, just like `sizeof(AM_LONG)`, but
the values can differ by platform (Windows with newer Perl was failing
intermittently).

Fixes #60.