Remove variable pre-declaration
That hasn't been required since C99... when I was in 6th grade. That was
when the Perl version was 5.6.0, which we definitely don't support.

Since we don't need to pre-declare variables everywhere, we also don't
need the XS `PREINIT` sections anymore. Another magical XS thingy gone!
I'm very happy with that.