Adjust for COW-shared constant-folded PV buffers (Perl v5.41)
https://github.com/Perl/perl5/commit/06e421c559c63975f29c35ba3588a0e6b0c75eca
made a change which allows constant folded PVs to have the `IsCOW` flag. The
aim of this is to reduce unnecessary PV buffer copying and lower memory usage.

This caused some breakage in `AM.xs`, as some `sum[]` PV buffers became shared
following this change, making it unsafe to modify them as-is. This commit calls
`SV_CHECK_THINKFIRST()` on SVs found to be affected, which un-COWs the buffer by
taking a fresh copy of it, prior to modifications being made.